refactor: :recycle: UserSearchBar view, logic 분할 (예시)
